Slight refactor
