+++ src/components/FAQLayout/index.tsx
@@ -3,16 +3,12 @@ import FAQSection from "../FAQSection"
 
 interface Props {
   title: string
-  permalink: string
-  questions: string
   rawMarkdownBody: string
   pageContext: any
 }
 
 const index = ({
   title,
-  permalink,
-  questions,
   rawMarkdownBody,
   pageContext
 }: Props) => {
@@ -21,8 +17,6 @@ const index = ({
       <div className="documentationContent">
         <FAQSection
           title={title}
-          permalink={permalink}
-          questions={questions}
           rawMarkdownBody={rawMarkdownBody}
           pageContext={pageContext}
         />
