+++ src/content/faq/Frontend.md
@@ -2,7 +2,6 @@
 title: Frontend
 layout: faq
 permalink: /faq/#frontend
-questions: Does GraphQL replace Redux or other state management libraries?
 position: 5
 ---
