+++ src/content/faq/Specification.md
@@ -2,7 +2,6 @@
 title: Specification
 layout: faq
 permalink: /faq/#specification
-questions: What is the best way to follow specification releases?,How can I contribute to the GraphQL specification?,Where is the documentation for subscriptions?
 position: 4
 ---
